requirements.txt
asgiref==3.7.2
Django==4.2.6
Pillow==10.1.0
sqlparse==0.4.4
typing-extensions==4.8.0
django-filter==23.5 
djangorestframework==3.14.0
Markdown==3.5.2 
